code clean, merged NodeDefWithoutCyclesDeleteJob, RecordsRemovedCyclesDeleteJob, UserPrefsCycleResetJob into CyclesDeletedCheckJob
