+++ common/i18n/resources/en.js
@@ -468,12 +468,11 @@ module.exports = {
     CategoriesValidationJob: 'Categories Validation',
     CollectImportJob: 'Collect Import',
     CollectSurveyReaderJob: 'Collect Survey Reader',
+    CyclesDeletedCheckJob: 'Check deleted cycles',
     NodeDefsImportJob: 'Node Defs Import',
-    NodeDefWithoutCyclesDeleteJob: 'Delete Node Definitions without cycles',
     NodeDefsValidationJob: 'Node Defs Validation',
     RecordCheckJob: 'Record Check',
     RecordsImportJob: 'Records Import',
-    RecordsRemovedCyclesDeleteJob: 'Delete records of removed cycles',
     RecordsUniquenessValidationJob: 'Records Uniqueness Validation',
     SamplingPointDataImportJob: 'Sampling Point Data Import',
     SurveyCreatorJob: 'Survey Create',
@@ -487,7 +486,6 @@ module.exports = {
     TaxonomiesImportJob: 'Taxonomies Import',
     TaxonomiesValidationJob: 'Taxonomies Validation',
     TaxonomyImportJob: 'Taxonomy Import',
-    UserPrefsCycleResetJob: 'Reset deleted cycles user prefs',
   },
 
   // ====== App Errors
