+++ server/modules/survey/service/publish/surveyPublishJob.js
@@ -4,11 +4,9 @@ const NodeDefsValidationJob = require('./jobs/nodeDefsValidationJob')
 const CategoriesValidationJob = require('./jobs/categoriesValidationJob')
 const TaxonomiesValidationJob = require('./jobs/taxonomiesValidationJob')
 const SurveyInfoValidationJob = require('./jobs/surveyInfoValidationJob')
-const NodeDefWithoutCyclesDeleteJob = require('./jobs/nodeDefWithoutCyclesDeleteJob')
-const RecordsRemovedCyclesDeleteJob = require('./jobs/recordsRemovedCyclesDeleteJob')
 const RecordCheckJob = require('../recordCheckJob')
 const SurveyPropsPublishJob = require('./jobs/surveyPropsPublishJob')
-const UserPrefsCycleResetJob = require('./jobs/userPrefsCycleResetJob')
+const CyclesDeletedCheckJob = require('./jobs/cyclesDeletedCheckJob')
 const SurveyDependencyGraphsGenerationJob = require('../surveyDependencyGraphsGenerationJob')
 const SurveyRdbGeneratorJob = require('../../../surveyRdb/service/surveyRdbGeneratorJob')
 class SurveyPublishJob extends Job {
@@ -19,11 +17,9 @@ class SurveyPublishJob extends Job {
       new CategoriesValidationJob(),
       new TaxonomiesValidationJob(),
       new SurveyInfoValidationJob(),
-      new NodeDefWithoutCyclesDeleteJob(),
-      new RecordsRemovedCyclesDeleteJob(),
+      new CyclesDeletedCheckJob(),
       // record check must be executed before publishing survey props
       new RecordCheckJob(),
-      new UserPrefsCycleResetJob(),
       new SurveyPropsPublishJob(),
       new SurveyDependencyGraphsGenerationJob(),
       new SurveyRdbGeneratorJob(),
